added geocoding functionality and attached to view
