--- public/modules/core/css/.form/form-checkbox.css
@@ -1,117 +1,0 @@
-/* basic css */
-.paper-checkbox  {
-    position: relative;
-}
-.paper-checkbox > input[type=checkbox] {
-    display: none;
-}
-.paper-checkbox > label,
-.paper-checkbox p {
-    padding-left: 50px;
-}
-
-/* checkbox css */
-
-.paper-checkbox > label {
-    cursor: pointer;
-}
-.paper-checkbox label > span {
-    display: block;
-    position: absolute;
-    left: 0;
-    -webkit-transition-duration: 0.3s;
-    -moz-transition-duration: 0.3s;
-    transition-duration: 0.3s;
-}
-.paper-checkbox label > .circle{
-    background-color: transparent;
-    left: -30px;
-    top: -30px;
-    height: 80px;
-    width: 80px;
-    /* z-index: -1; */
-    border-radius: 50%;
-    -webkit-animation: growCircle 0.3s ease;
-    -moz-animation: growCircle 0.3s ease;
-    animation: growCircle 0.3s ease;
-}
-.paper-checkbox label > .box {
-    border: 2px solid #000;
-    height: 20px;
-    width: 20px;
-    z-index: 888;
-    -webkit-transition-delay: 0.2s;
-    -moz-transition-delay: 0.2s;
-    transition-delay: 0.2s;
-}
-.paper-checkbox label > .check {
-    top: -7px;
-    left: 6px;
-    width: 12px;
-    height: 24px;
-    border: 2px solid #0f9d58;
-    border-top: none;
-    border-left: none;
-    opacity: 0;
-    z-index: 888;
-    -webkit-transform: rotate(180deg);
-    -moz-transform: rotate(180deg);
-    transform: rotate(180deg);
-    -webkit-transition-delay: 0.3s;
-    -moz-transition-delay: 0.3s;
-    transition-delay: 0.3s;
-}
-/* handling click events */
-
-/* when checkbox is checked */
-
-
-.paper-checkbox input[type=checkbox]:checked ~ label > .box {
-    opacity: 0;
-    -webkit-transform: scale(0) rotate(-180deg);
-    -moz-transform: scale(0) rotate(-180deg);
-    transform: scale(0) rotate(-180deg);
-}
-.paper-checkbox input[type=checkbox]:checked ~ label > .check {
-    opacity: 1;
-    -webkit-transform: scale(1) rotate(45deg);
-    -moz-transform: scale(1) rotate(45deg);
-    transform: scale(1) rotate(45deg);
-}
-/* bubble animation */
-
-@-webkit-keyframes growCircle {
-    0%,
-    100% {
-        -webkit-transform: scale(0);
-        transform: scale(0);
-    }
-    70% {
-        opacity: 0.5;
-        background: #ddd;
-        -webkit-transform: scale(1.25);
-        transform: scale(1.25);
-    }
-}
-@-moz-keyframes growCircle {
-    0%,
-    100% {
-        -moz-transform: scale(0);
-    }
-    70% {
-        opacity: 0.5;
-        background: #DDD;
-        -moz-transform: scale(1.25);
-    }
-}
-@keyframes growCircle {
-    0%,
-    100% {
-        transform: scale(0);
-    }
-    70% {
-        opacity: 0.5;
-        background: #DDD;
-        transform: scale(1.25);
-    }
-}
