--- public/modules/core/css/.form/form-style.css
@@ -1,30 +1,0 @@
-body {
-    background-image:url(http://scotch.io/wp-content/uploads/2014/07/61.jpg);
-}
-
-.container {
-    font-family:'Roboto';
-    width:600px;
-    margin:30px auto 0;
-    display:block;
-    background:#FFF;
-    padding:10px 50px 50px;
-}
-
-h2 {
-    text-align:center;
-    margin-bottom:50px;
-}
-
-h2 > small {
-    font-weight:normal;
-    color:#888;
-    display:block;
-}
-.footer {
-    text-align:center;
-}
-
-.footer a {
-    color:#53B2C8;
-}
