--- public/modules/form/form.client.module.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// Use applicaion configuration module to register a n`ew module
-ApplicationConfiguration.registerModule('project-submission');
