--- public/modules/form/views/project-submit-wiz-client.view.html
@@ -1,33 +1,0 @@
-<section data-ng-controller="formController">
-
-    <div class="row">
-
-        <div class="col-sm-8 col-sm-offset-2">
-            <div id="form-container">
-
-                <div class="page-header text-center">
-                    <h2>Mapping Salt Lake City</h2>
-                    <p>Thanks for your dedication to the community.</p>
-                    </br>
-
-                    <!-- the links to our nested states using relative paths -->
-                    <!-- add the active class if the state matches our ui-sref -->
-                    <div id="status-buttons" class="text-center">
-                        <a ui-sref-active="active" ui-sref=".page1"><span>1</span> The Basics</a>
-                        <a ui-sref-active="active" ui-sref=".page2"><span>2</span> Some Details</a>
-                        <a ui-sref-active="active" ui-sref=".page3"><span>3</span> Your Project</a>
-                    </div>
-
-                </div>
-
-                <form id="signup-form" ng-submit="createProject()">
-                    <!-- nested state views will be injected here -->
-                    <div id="form-views" ui-view></div>
-                </form>
-
-            </div>
-        </div>
-
-    </div>
-
-</section>
