--- public/modules/form/views/project-submit-wiz-index.client.view.html
@@ -1,9 +1,0 @@
-<section data-ng-controller="formController">
-
-    <!-- views will be injected here -->
-
-    <div class="container">
-        <div ui-view></div>
-    </div>
-
-</section>
