--- public/modules/form/views/project-submit-wiz-page1-client.view.html
@@ -1,43 +1,0 @@
-<section data-ng-controller="formController">
-
-
-    <div class="form-group">
-
-
-        <div class="form-group col-md-6">
-            <input type="text" data-ng-model="firstName" id="firstName" class="form-control" name="firstName"
-                   placeholder="First Name" required>
-        </div>
-
-        <div class="form-group">
-            <input type="text" data-ng-model="lastName" id="lastName" class="form-control" name="lastName"
-                   placeholder="Last Name" required>
-        </div>
-
-        <div class="form-group">
-            <input type="email" data-ng-model="email" id="email" class="form-control" name="email"
-                   placeholder="E-mail" required>
-        </div>
-
-        <div class="form-group">
-            <input type="password" data-ng-model="password" id="password" class="form-control" name="password"
-                   placeholder="Password" required>
-        </div>
-
-        <div class="form-group">
-            <input type="password" class="form-control" name="password"
-                   placeholder="Confirm Password" required>
-        </div>
-
-
-
-    </div>
-
-    <div class="form-group row">
-        <div class="col-xs-6 col-xs-offset-6">
-            <a ui-sref="signupwiz.page2" class="btn btn-block btn-info">
-                Next  <span class="glyphicon glyphicon-circle-arrow-right"></span>
-            </a>
-        </div>
-    </div>
-</section>
