--- public/modules/form/views/project-submit-wiz-page2-client.view.html
@@ -1,45 +1,0 @@
-<section data-ng-controller="formController">
-    <label>Some Details</label>
-
-    <div class="form-group">
-        <div class="col-xs-10">
-            <div class="form-inline">
-
-                <div class="form-group col-md-6">
-                    <input type="text" data-ng-model="projectTitle" id="projectTitle" class="form-control" name="projectTitle"
-                           placeholder="The Title of Your Project" required>
-                </div>
-                <div class="text">
-                    <input type="text" class="form-control" name="projectStreet" ng-model="projectStreet" id="projectStreet" placeholder="Your Project's Street Location*">
-                </div>
-                <div class="text">
-                    <input type="number" class="form-control" name="projectZip" ng-model="projectZip" id="projectZip" placeholder="Zip Code*">
-                </div>
-
-                <p>This adddress and zip code information will be used to place a link to your project on the home page map. We understand many projects either have more than one location or do not have a precise location. In these cases, please come up with an address and zip code that makes the most sense to you.</p>
-
-
-            </div>
-        </div>
-
-        <div class="form-group row">
-            <div class="col-xs-5">
-                <a ui-sref="signupwiz.page1" class="btn btn-block btn-info">
-                    <span class="glyphicon glyphicon-circle-arrow-left"></span> Back
-                </a>
-            </div>
-
-
-            <div class="col-xs-5 col-xs-offset-6">
-                <a ui-sref="signupwiz.page3" class="btn btn-block btn-info">
-                    Next <span class="glyphicon glyphicon-circle-arrow-right"></span>
-                </a>
-            </div>
-
-        </div>
-
-
-
-
-
-</section>
