+++ public/modules/map/views/map.client.view.html
@@ -1,6 +1,6 @@
 <section class="container" data-ng-controller="MapController" data-ng-init="mapFunction()">
 
-    <div id="logo" >
+    <div id="logo" class="img-responsive">
         <!--logo goes here -- from map.css-->
     </div>
 
@@ -64,7 +64,7 @@
     </span>
 
     <!--to call MapBox map-->
-    <!--turned off to save map views-->
+
     <div id="map" class="map"></div>
 
 </section>
