+++ public/modules/projects/views/edit-project.client.view.html
@@ -56,9 +56,6 @@
                     <!--</div>-->
                 </div>
                 <div class="form-group">
-                    <button class="btn btn-primary btn-lg btn-main grow" data-ng-click="geocode(); mapImage()">Geocode Me, Baby!</button>
-                </div>
-                <div class="form-group">
                     <input type="submit" value="Update" class="btn btn-primary btn-lg btn-main grow">
                 </div>
 				<div data-ng-show="error" class="text-danger">
